Split client restriction supportability
